feat: Migrate API routes to Drizzle ORM
- Rename route.drizzle.ts files to route.ts for active use
- Keep legacy versions as route.legacy.ts for reference
- Migrated routes:
  - /api/adjustments - Main adjustments API
  - /api/user/preferences - User preferences API
  - /api/user/preferences/init - Preferences initialization

This completes the initial migration of auth-related routes to Drizzle ORM.
